+++ packages/babel-plugin/src/visitor.ts
@@ -36,7 +36,7 @@ export const visitor = ({ types: t, template }: Core) => {
         // Process CSS function calls and generate the hashed classNames
         // processCSS(path, t, template, importedStyleFunctions);
         // Process TaggedTemplateExpressions with styled components and generate the hashed classNames
-        processTaggedTemplateExpression(path, template, importedStyleFunctions);
+        // processTaggedTemplateExpression(path, template, importedStyleFunctions);
         // Traversal over the JSX elements in the Program node to identify Kuma-UI components,
         // processComponents(path, importedStyleFunctions);