+++ packages/compiler/src/compile.ts
@@ -9,6 +9,7 @@ import { collectPropsFromJsx } from "./collector";
 import { extractProps } from "./extractor";
 import { componentList } from "@kuma-ui/core/components/componentList";
 import { optimize } from "./optimizer/optimize";
+import { processTaggedTemplateExpression } from "./processTaggedTemplateExpression";
 
 const project = new Project({});
 
@@ -36,9 +37,11 @@ const compile = (
       const originalComponentName = Object.keys(bindings).find(
         (key) =>
           bindings[key] === jsxTagName &&
+          // eslint-disable-next-line @typescript-eslint/no-unsafe-argument -- FIXME
           Object.values(componentList).some((c) => c === key)
       );
       if (!originalComponentName) return;
+      // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment -- FIXME
       const componentName =
         originalComponentName as (typeof componentList)[keyof typeof componentList];
       const extractedPropsMap = collectPropsFromJsx(openingElement);
@@ -50,6 +53,9 @@ const compile = (
       if (result) css.push(result.css);
 
       optimize(componentName, openingElement, extractedPropsMap["as"]);
+    }
+    if (Node.isTaggedTemplateExpression(node)) {
+      processTaggedTemplateExpression(node, bindings);
     }
   });
   return { code: source.getFullText(), id, css: css.join(" ") };