+++ website/src/components/example/Image.tsx
@@ -3,7 +3,7 @@ import { Image } from "@kuma-ui/core";
 export const ImageExample = () => {
   return (
     <Image
-      src="https://raw.githubusercontent.com/poteboy/kuma-ui/main/media/logo.webp"
+      src="https://raw.githubusercontent.com/kuma-ui/kuma-ui/main/media/logo.webp"
       alt="Description of image"
     />
   );