+++ packages/next-plugin/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @kuma-ui/next-plugin
 
+## 1.1.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @kuma-ui/core@1.1.1
+  - @kuma-ui/webpack-plugin@1.1.3
+
+## 1.1.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @kuma-ui/webpack-plugin@1.1.2
+
 ## 1.1.1
 
 ### Patch Changes