+++ packages/vite/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @kuma-ui/vite
 
+## 1.0.6
+
+### Patch Changes
+
+- Updated dependencies [[`8a42a22`](https://github.com/kuma-ui/kuma-ui/commit/8a42a225359bad6d2bf35a3c63d897ae02f59392)]:
+  - @kuma-ui/system@1.1.0
+  - @kuma-ui/babel-plugin@1.1.2
+
+## 1.0.5
+
+### Patch Changes
+
+- Updated dependencies [[`439965c`](https://github.com/kuma-ui/kuma-ui/commit/439965c78ff2d1a6e2f4ac9225ad9a9067d36ee2)]:
+  - @kuma-ui/babel-plugin@1.1.1
+
 ## 1.0.4
 
 ### Patch Changes